Add files via upload
The following documents is about a data set from a mechanic the goal was: 
Use the data to create meaningful information.  Answer the following questions:  1. Trevor's best customers or top 10 in terms of volume and revenue, 2. Trevor's worst or bottom 10 in terms of lowest volume and lowest revenue and 3. the mechanics that perform the most work and what type of repairs for customers are performed more frequently.  In addition, consider what other information or data that Trevor needs or could he use to help his business to improve?